split ui and logic
